Update create_training.py
We updated the vame.create_trainset() function with more flexibility in terms of checking the parameter used to preprocess the data.
By flagging the parameter check_parameter=True, the function will take only a fraction of the data and plot the original as well as filtered data to compare. 
Moreover, by setting the parameter fixed=True, the function recognizes that the data comes from a headfixed setup, i.e. created from csv_to_numpy